chore: reorganize types into pokemonApi.type and types folder
