+++ src/api/pokemonApi.types.ts
@@ -37,13 +37,3 @@ export interface PokemonAPIResponse {
   height: number;
   weight: number;
 }
-
-export interface Pokemon {
-  id: number;
-  name: string;
-  image: string;
-  types: string[];
-  abilities: string[];
-  height: number;
-  weight: number;
-}
