+++ src/components/Theme/index.ts
@@ -1,0 +1,4 @@
+import ThemeProvider from './ThemeProvider';
+import ThemeToggle from './ThemeToggle';
+
+export { ThemeProvider, ThemeToggle };
