+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,9 +1,5 @@
 const NotFoundPage = () => {
-  return (
-    <>
-      <h2>Not found page. Error: 404 </h2>
-    </>
-  );
+  return <h2>Not found page. Error: 404 </h2>;
 };
 
 export default NotFoundPage;
