+++ src/types/pokemon.ts
@@ -1,0 +1,9 @@
+export interface Pokemon {
+  id: number;
+  name: string;
+  image: string;
+  types: string[];
+  abilities: string[];
+  height: number;
+  weight: number;
+}
